ai_powered/requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
spacy>=3.5.0
tqdm
sentencepiece
regex
datasets
torch
scikit-learn
accelerate>=0.26.0
streamlit>=1.24.0
pytest>=7.0.0
langchain 
langchain-groq 
groq
python-dotenv
langchain-community
llama-cpp-python --prefer-binary

